Add validation provider service
The validation in JSR303 needs to be given a Validation Provider which normally works using the classpath but since we are using OSGI this can be problematic.
In Geronimo the included implementation is already wired so Validation.buildDefaultValidatorFactory(); is enough, in other setups, like for example the unit tests, we have to specify the implementation hence ValidationServiceTestImpl which is hardcoded to find the apache bval implementation.

By using an OSGI service ValidationService we can then have other modules access the validation in both contexts .