Pass the validation errors through the REST response back to the web client
Here the data structure allows to know which fields failed, which bad value was given and what message the validator returned.
As an example the Note text is required to be not empty (actually here Size > 1), the client UI can extract and display the note text error next to the input field.